refactor: add socialLogin to remove repetitive code
